Refactor CustomPlanetView widget to use AppColors for progress bar color
Previously, the `CustomPlanetView` widget used a hardcoded `Colors.transparent` value for the `progressBarColor` property of the `Flutter3DViewer` widget. This commit introduces the use of the `AppColors.grey` from the `app_colors.dart` file to set the `progressBarColor`. This change aligns the widget with the application's theme and improves consistency across the UI.